chore: 전국 행정동 시드 SQL 추가 및 region_code BIGINT 전환

+++ src/main/java/com/sparta/member_regions_service/adaptor/in/web/vo/AddMemberRegionRequestVo.java
@@ -8,7 +8,7 @@ public record AddMemberRegionRequestVo(
 		// 지역 코드
 		@NotNull(message = "지역 코드는 필수입니다.")
 		@Positive(message = "지역 코드가 올바르지 않습니다.")
-		Integer regionCode,
+		Long regionCode,
 		// 대표 지정 여부 (선택)
 		Boolean primary
 ) {
