+++ src/main/java/com/sparta/member_regions_service/adaptor/in/web/vo/MemberRegionResponseVo.java
@@ -7,7 +7,7 @@ public record MemberRegionResponseVo(
 		Long memberRegionId,
 		String memberUuid,
 		boolean primary,
-		int regionCode,
+		long regionCode,
 		String regionName,
 		boolean verified,
 		Instant verifiedAt,
